Follow the same pattern for tests
